pyproject.toml
[project]
name = "diablo"
version = "0.1.0"
description = "AI-Powered Smart Contract Security Intelligence"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{name = "Shady"}]

dependencies = [
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "httpx>=0.27",
    "openai>=1.10",
    "python-dotenv>=1.0",
    "pydantic>=2.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["backend*"]
exclude = ["extension*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
